test: ignore resource group destroy
