+++ tests/pr_test.go
@@ -74,6 +74,10 @@ func setupOptionsQuickStartPattern(t *testing.T, prefix string, dir string) *tes
 			"ssh_key": sshPublicKey,
 		},
 		CloudInfoService: sharedInfoSvc,
+		ImplicitRequired: true,
+		ImplicitDestroy: []string{
+			"module.landing_zone.module.landing_zone.ibm_resource_group.resource_groups",
+		},
 	})
 
 	return options
@@ -140,6 +144,10 @@ func setupOptionsRoksPattern(t *testing.T, prefix string) *testhelper.TestOption
 		Prefix:           prefix,
 		ResourceGroup:    resourceGroup,
 		CloudInfoService: sharedInfoSvc,
+		ImplicitRequired: true,
+		ImplicitDestroy: []string{
+			"module.roks_landing_zone.module.landing_zone.ibm_resource_group.resource_groups",
+		},
 	})
 
 	options.TerraformVars = map[string]interface{}{
@@ -173,6 +181,10 @@ func setupOptionsVsiPattern(t *testing.T, prefix string) *testhelper.TestOptions
 		Prefix:           prefix,
 		ResourceGroup:    resourceGroup,
 		CloudInfoService: sharedInfoSvc,
+		ImplicitRequired: true,
+		ImplicitDestroy: []string{
+			"module.vsi_landing_zone.module.landing_zone.ibm_resource_group.resource_groups",
+		},
 	})
 
 	options.TerraformVars = map[string]interface{}{
@@ -206,6 +218,10 @@ func setupOptionsVpcPattern(t *testing.T, prefix string) *testhelper.TestOptions
 		Prefix:           prefix,
 		ResourceGroup:    resourceGroup,
 		CloudInfoService: sharedInfoSvc,
+		ImplicitRequired: true,
+		ImplicitDestroy: []string{
+			"module.vpc_landing_zone.module.landing_zone.ibm_resource_group.resource_groups",
+		},
 	})
 
 	options.TerraformVars = map[string]interface{}{
@@ -296,78 +312,6 @@ func setupOptionsSchematics(t *testing.T, prefix string, dir string) *testschema
 	})
 
 	return options
-}
-
-func TestRunVSIQuickStartPatternSchematics(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptionsSchematics(t, "qs-sc", quickStartPatternTerraformDir)
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "region", Value: options.Region, DataType: "string"},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "ssh_key", Value: sshPublicKey(t), DataType: "string"},
-		{Name: "service_endpoints", Value: "private", DataType: "string"},
-	}
-
-	err := options.RunSchematicTest()
-	assert.NoError(t, err, "Schematic Test had unexpected error")
-}
-
-func TestRunVSIPatternSchematics(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptionsSchematics(t, "vsi-sc", vsiPatternTerraformDir)
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "region", Value: options.Region, DataType: "string"},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "ssh_public_key", Value: sshPublicKey(t), DataType: "string"},
-		{Name: "add_atracker_route", Value: add_atracker_route, DataType: "bool"},
-		{Name: "service_endpoints", Value: "private", DataType: "string"},
-	}
-
-	err := options.RunSchematicTest()
-	assert.NoError(t, err, "Schematic Test had unexpected error")
-}
-
-func TestRunRoksPatternSchematics(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptionsSchematics(t, "ocp-sc", roksPatternTerraformDir)
-
-	options.WaitJobCompleteMinutes = 120
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "region", Value: options.Region, DataType: "string"},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "tags", Value: options.Tags, DataType: "list(string)"},
-		{Name: "service_endpoints", Value: "private", DataType: "string"},
-	}
-
-	err := options.RunSchematicTest()
-	assert.NoError(t, err, "Schematic Test had unexpected error")
-}
-
-func TestRunVPCPatternSchematics(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptionsSchematics(t, "vpc-sc", vpcPatternTerraformDir)
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "region", Value: options.Region, DataType: "string"},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "tags", Value: options.Tags, DataType: "list(string)"},
-		{Name: "add_atracker_route", Value: add_atracker_route, DataType: "bool"},
-		{Name: "service_endpoints", Value: "private", DataType: "string"},
-	}
-
-	err := options.RunSchematicTest()
-	assert.NoError(t, err, "Schematic Test had unexpected error")
 }
 
 func TestRunVsiExtention(t *testing.T) {
@@ -475,6 +419,8 @@ func TestRunVsiExtention(t *testing.T) {
 		fmt.Println("Terratest failed. Debug the test and delete resources manually.")
 	} else {
 		logger.Log(t, "START: Destroy (existing resources)")
+		// ignore resource groups when destroying
+		terraform.RunTerraformCommand(t, existingTerraformOptions, "state", "rm", "module.vpc_landing_zone.module.landing_zone.ibm_resource_group.resource_groups")
 		terraform.Destroy(t, existingTerraformOptions)
 		terraform.WorkspaceDelete(t, existingTerraformOptions, prefix)
 		logger.Log(t, "END: Destroy (existing resources)")
